requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.2
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql==0.0.3
mysql-connector-python==8.3.0
mysqlclient==2.2.4
setuptools==65.5.1
Werkzeug==3.0.1
wheel==0.38.4
zipp==3.17.0
python-dotenv==1.0.1